# ENH: stats.permutation_test: improve performance of samples/pairings random permutations