--- a/scipy/stats/_resampling.py
+++ b/scipy/stats/_resampling.py
@@ -895,10 +895,12 @@ def _batch_generator(iterable, batch):
 
 def _pairings_permutations_gen(n_permutations, n_samples, n_obs_sample, batch,
                                random_state):
-
+    # Returns a generator that yields arrays of size
+    # `(batch, n_samples, n_obs_sample)`.
+    # Each row is an independent permutation of indices 0 to `n_obs_sample`.
     batch = min(batch, n_permutations)
 
-    if isinstance(random_state, np.random.Generator):
+    if hasattr(random_state, 'permuted'):
         def batched_perm_generator():
             indices = np.arange(n_obs_sample)
             indices = np.tile(indices, (batch, n_samples, 1))
@@ -907,7 +909,7 @@ def _pairings_permutations_gen(n_permutations, n_samples, n_obs_sample, batch,
                 # Don't permute in place, otherwise results depend on `batch`
                 permuted_indices = random_state.permuted(indices, axis=-1)
                 yield permuted_indices[:batch_actual]
-    else:
+    else:  # RandomState and early Generators don't have `permuted`
         def batched_perm_generator():
             for k in range(0, n_permutations, batch):
                 batch_actual = min(batch, n_permutations-k)