--- a/scipy/stats/tests/test_resampling.py
+++ b/scipy/stats/tests/test_resampling.py
@@ -903,6 +903,9 @@ class TestPermutationTest:
 
     rtol = 1e-14
 
+    def setup_method(self):
+        self.rng = np.random.default_rng(7170559330470561044)
+
     # -- Input validation -- #
 
     def test_permutation_test_iv(self):
@@ -966,9 +969,8 @@ class TestPermutationTest:
     def test_batch(self, permutation_type, random_state):
         # make sure that the `batch` parameter is respected by checking the
         # maximum batch size provided in calls to `statistic`
-        np.random.seed(0)
-        x = np.random.rand(10)
-        y = np.random.rand(10)
+        x = self.rng.random(10)
+        y = self.rng.random(10)
 
         def statistic(x, y, axis):
             batch_size = 1 if x.ndim == 1 else len(x)
@@ -1009,9 +1011,8 @@ class TestPermutationTest:
     def test_permutations(self, permutation_type, exact_size, random_state):
         # make sure that the `permutations` parameter is respected by checking
         # the size of the null distribution
-        np.random.seed(0)
-        x = np.random.rand(3)
-        y = np.random.rand(3)
+        x = self.rng.random(3)
+        y = self.rng.random(3)
 
         def statistic(x, y, axis):
             return np.mean(x, axis=axis) - np.mean(y, axis=axis)
@@ -1088,11 +1089,7 @@ class TestPermutationTest:
         # check that the randomized and exact tests agree to reasonable
         # precision for permutation_type='pairings'
 
-        alternative = 'two-sided'
-        try:
-            rng = np.random.default_rng(1)
-        except AttributeError:
-            rng = np.random.RandomState(1)
+        alternative, rng = 'two-sided', self.rng
 
         nx, ny, permutations = 8, 8, 40000
         assert special.factorial(nx) > permutations
@@ -1128,16 +1125,18 @@ class TestPermutationTest:
         x = np.arange(3*4*5).reshape(3, 4, 5)
         y = np.moveaxis(np.arange(4)[:, None, None], 0, axis)
 
+        rng1 = np.random.default_rng(4337234444626115331)
         res1 = stats.ttest_ind(x, y, permutations=permutations, axis=axis,
-                               random_state=0, alternative=alternative)
+                               random_state=rng1, alternative=alternative)
 
         def statistic(x, y, axis):
             return stats.ttest_ind(x, y, axis=axis).statistic
 
+        rng2 = np.random.default_rng(4337234444626115331)
         res2 = permutation_test((x, y), statistic, vectorized=True,
                                 n_resamples=permutations,
                                 alternative=alternative, axis=axis,
-                                random_state=0)
+                                random_state=rng2)
 
         assert_allclose(res1.statistic, res2.statistic, rtol=self.rtol)
         assert_allclose(res1.pvalue, res2.pvalue, rtol=self.rtol)
@@ -1146,9 +1145,9 @@ class TestPermutationTest:
 
     @pytest.mark.parametrize('alternative', ("less", "greater", "two-sided"))
     def test_against_ks_2samp(self, alternative):
-        rng = np.random.default_rng(abs(hash("ks_2samp")))
-        x = rng.normal(size=4, scale=1)
-        y = rng.normal(size=5, loc=3, scale=3)
+
+        x = self.rng.normal(size=4, scale=1)
+        y = self.rng.normal(size=5, loc=3, scale=3)
 
         expected = stats.ks_2samp(x, y, alternative=alternative, mode='exact')
 
@@ -1159,16 +1158,16 @@ class TestPermutationTest:
         # ks_2samp is always a one-tailed 'greater' test
         # it's the statistic that changes (D+ vs D- vs max(D+, D-))
         res = permutation_test((x, y), statistic1d, n_resamples=np.inf,
-                               alternative='greater')
+                               alternative='greater', random_state=self.rng)
 
         assert_allclose(res.statistic, expected.statistic, rtol=self.rtol)
         assert_allclose(res.pvalue, expected.pvalue, rtol=self.rtol)
 
     @pytest.mark.parametrize('alternative', ("less", "greater", "two-sided"))
     def test_against_ansari(self, alternative):
-        np.random.seed(0)
-        x = stats.norm.rvs(size=4, scale=1)
-        y = stats.norm.rvs(size=5, scale=3)
+
+        x = self.rng.normal(size=4, scale=1)
+        y = self.rng.normal(size=5, scale=3)
 
         # ansari has a different convention for 'alternative'
         alternative_correspondence = {"less": "greater",
@@ -1181,16 +1180,16 @@ class TestPermutationTest:
             return stats.ansari(x, y).statistic
 
         res = permutation_test((x, y), statistic1d, n_resamples=np.inf,
-                               alternative=alternative)
+                               alternative=alternative, random_state=self.rng)
 
         assert_allclose(res.statistic, expected.statistic, rtol=self.rtol)
         assert_allclose(res.pvalue, expected.pvalue, rtol=self.rtol)
 
     @pytest.mark.parametrize('alternative', ("less", "greater", "two-sided"))
     def test_against_mannwhitneyu(self, alternative):
-        np.random.seed(0)
-        x = stats.uniform.rvs(size=(3, 5, 2), loc=0)
-        y = stats.uniform.rvs(size=(3, 5, 2), loc=0.05)
+
+        x = stats.uniform.rvs(size=(3, 5, 2), loc=0, random_state=self.rng)
+        y = stats.uniform.rvs(size=(3, 5, 2), loc=0.05, random_state=self.rng)
 
         expected = stats.mannwhitneyu(x, y, axis=1, alternative=alternative)
 
@@ -1199,15 +1198,15 @@ class TestPermutationTest:
 
         res = permutation_test((x, y), statistic, vectorized=True,
                                n_resamples=np.inf, alternative=alternative,
-                               axis=1)
+                               axis=1, random_state=self.rng)
 
         assert_allclose(res.statistic, expected.statistic, rtol=self.rtol)
         assert_allclose(res.pvalue, expected.pvalue, rtol=self.rtol)
 
     def test_against_cvm(self):
-        np.random.seed(0)
-        x = stats.norm.rvs(size=4, scale=1)
-        y = stats.norm.rvs(size=5, loc=3, scale=3)
+
+        x = stats.norm.rvs(size=4, scale=1, random_state=self.rng)
+        y = stats.norm.rvs(size=5, loc=3, scale=3, random_state=self.rng)
 
         expected = stats.cramervonmises_2samp(x, y, method='exact')
 
@@ -1217,7 +1216,7 @@ class TestPermutationTest:
 
         # cramervonmises_2samp has only one alternative, greater
         res = permutation_test((x, y), statistic1d, n_resamples=np.inf,
-                               alternative='greater')
+                               alternative='greater', random_state=self.rng)
 
         assert_allclose(res.statistic, expected.statistic, rtol=self.rtol)
         assert_allclose(res.pvalue, expected.pvalue, rtol=self.rtol)
@@ -1231,12 +1230,12 @@ class TestPermutationTest:
         # and random permutation tests approximate SciPy's asymptotic pvalues
         # and that exact and random permutation test results are even closer
         # to one another (than they are to the asymptotic results).
-        np.random.seed(0)
 
         # Three samples, different (but compatible) shapes with different ndims
-        x = np.random.rand(3)
-        y = np.random.rand(1, 3, 2)
-        z = np.random.rand(2, 1, 4)
+        rng = np.random.default_rng(6709265303529651545)
+        x = rng.random(size=(3))
+        y = rng.random(size=(1, 3, 2))
+        z = rng.random(size=(2, 1, 4))
         data = (x, y, z)
 
         # Define the statistic (and pvalue for comparison)
@@ -1258,7 +1257,7 @@ class TestPermutationTest:
 
         # Calculate exact and randomized permutation results
         kwds = {'vectorized': False, 'axis': axis, 'alternative': 'greater',
-                'permutation_type': 'independent', 'random_state': 0}
+                'permutation_type': 'independent', 'random_state': self.rng}
         res = permutation_test(data, statistic1d, n_resamples=np.inf, **kwds)
         res2 = permutation_test(data, statistic1d, n_resamples=1000, **kwds)
 
@@ -1272,9 +1271,9 @@ class TestPermutationTest:
 
     @pytest.mark.parametrize('alternative', ("less", "greater", "two-sided"))
     def test_against_wilcoxon(self, alternative):
-        np.random.seed(0)
-        x = stats.uniform.rvs(size=(3, 6, 2), loc=0)
-        y = stats.uniform.rvs(size=(3, 6, 2), loc=0.05)
+
+        x = stats.uniform.rvs(size=(3, 6, 2), loc=0, random_state=self.rng)
+        y = stats.uniform.rvs(size=(3, 6, 2), loc=0.05, random_state=self.rng)
 
         # We'll check both 1- and 2-sample versions of the same test;
         # we expect identical results to wilcoxon in all cases.
@@ -1294,13 +1293,11 @@ class TestPermutationTest:
         expected_stat = expected[0]
         expected_p = expected[1]
 
-        res1 = permutation_test((x-y,), statistic_1samp_1d, vectorized=False,
-                                permutation_type='samples', n_resamples=np.inf,
-                                alternative=alternative, axis=1)
-
-        res2 = permutation_test((x, y), statistic_2samp_1d, vectorized=False,
-                                permutation_type='samples', n_resamples=np.inf,
-                                alternative=alternative, axis=1)
+        kwds = {'vectorized': False, 'axis': 1, 'alternative': alternative,
+                'permutation_type': 'samples', 'random_state': self.rng,
+                'n_resamples': np.inf}
+        res1 = permutation_test((x-y,), statistic_1samp_1d, **kwds)
+        res2 = permutation_test((x, y), statistic_2samp_1d, **kwds)
 
         # `wilcoxon` returns a different statistic with 'two-sided'
         assert_allclose(res1.statistic, res2.statistic, rtol=self.rtol)
@@ -1312,8 +1309,8 @@ class TestPermutationTest:
 
     @pytest.mark.parametrize('alternative', ("less", "greater", "two-sided"))
     def test_against_binomtest(self, alternative):
-        np.random.seed(0)
-        x = np.random.randint(0, 2, size=10)
+
+        x = self.rng.integers(0, 2, size=10)
         x[x == 0] = -1
         # More naturally, the test would flip elements between 0 and one.
         # However, permutation_test will flip the _signs_ of the elements.
@@ -1327,16 +1324,16 @@ class TestPermutationTest:
 
         res = stats.permutation_test((x,), statistic, vectorized=True,
                                      permutation_type='samples',
-                                     n_resamples=np.inf,
+                                     n_resamples=np.inf, random_state=self.rng,
                                      alternative=alternative)
         assert_allclose(res.pvalue, expected.pvalue, rtol=self.rtol)
 
     # -- Exact Association Tests -- #
 
     def test_against_kendalltau(self):
-        np.random.seed(0)
-        x = stats.norm.rvs(size=6)
-        y = x + stats.norm.rvs(size=6)
+
+        x = self.rng.normal(size=6)
+        y = x + self.rng.normal(size=6)
 
         expected = stats.kendalltau(x, y, method='exact')
 
@@ -1345,7 +1342,7 @@ class TestPermutationTest:
 
         # kendalltau currently has only one alternative, two-sided
         res = permutation_test((x,), statistic1d, permutation_type='pairings',
-                               n_resamples=np.inf)
+                               n_resamples=np.inf, random_state=self.rng)
 
         assert_allclose(res.statistic, expected.statistic, rtol=self.rtol)
         assert_allclose(res.pvalue, expected.pvalue, rtol=self.rtol)
@@ -1356,14 +1353,15 @@ class TestPermutationTest:
         def statistic(x,):
             return np.sum((x == 1) & (y == 1))
 
-        np.random.seed(0)
         # x and y are binary random variables with some dependence
-        x = (np.random.rand(7) > 0.6).astype(float)
-        y = (np.random.rand(7) + 0.25*x > 0.6).astype(float)
+        rng = np.random.default_rng(6235696159000529929)
+        x = (rng.random(7) > 0.6).astype(float)
+        y = (rng.random(7) + 0.25*x > 0.6).astype(float)
         tab = stats.contingency.crosstab(x, y)[1]
 
         res = permutation_test((x,), statistic, permutation_type='pairings',
-                               n_resamples=np.inf, alternative=alternative)
+                               n_resamples=np.inf, alternative=alternative,
+                               random_state=rng)
         res2 = stats.fisher_exact(tab, alternative=alternative)
 
         assert_allclose(res.pvalue, res2[1])
@@ -1377,11 +1375,9 @@ class TestPermutationTest:
         # reproduces SciPy's exact pvalue and that random permutation test
         # approximates it.
 
-        np.random.seed(0)
-
-        x = np.random.rand(2, 4, 3)
-        y = np.random.rand(1, 4, 3)
-        z = np.random.rand(2, 4, 1)
+        x = self.rng.random(size=(2, 4, 3))
+        y = self.rng.random(size=(1, 4, 3))
+        z = self.rng.random(size=(2, 4, 1))
         x = stats.rankdata(x, axis=axis)
         y = stats.rankdata(y, axis=axis)
         z = stats.rankdata(z, axis=axis)
@@ -1402,6 +1398,7 @@ class TestPermutationTest:
         expected_statistic = statistic(*np.broadcast_arrays(*data), axis=axis)
         expected_pvalue = pvalue(*np.broadcast_arrays(*data), axis=axis)
 
+        # Let's forgive this use of an integer seed, please.
         kwds = {'vectorized': False, 'axis': axis, 'alternative': 'greater',
                 'permutation_type': 'pairings', 'random_state': 0}
         res = permutation_test(data, statistic1d, n_resamples=np.inf, **kwds)
@@ -1505,10 +1502,14 @@ class TestPermutationTest:
                2.6514917, 2.01160156, 0.47699563)
         cor.test(x, y, method = "spearm", alternative = "t")
         """
-        np.random.seed(0)
-        x = stats.norm.rvs(size=6)
-        y = x + stats.norm.rvs(size=6)
-
+        # data comes from
+        # np.random.seed(0)
+        # x = stats.norm.rvs(size=6)
+        # y = x + stats.norm.rvs(size=6)
+        x = [1.76405235, 0.40015721, 0.97873798,
+             2.2408932, 1.86755799, -0.97727788]
+        y = [2.71414076, 0.2488, 0.87551913,
+             2.6514917, 2.01160156, 0.47699563]
         expected_statistic = 0.7714285714285715
 
         def statistic1d(x):