#### Reference issue
gh-16962

#### What does this implement/fix?
gh-16962 brought up a case in which `permutation_test` was unreasonably slow. This is because `RandomState.permutation` doesn't operate on each axis independently, so we loop in pure Python to generate arrays of permutations. We can do better.

- Way back when I first ran into this, I found that `np.argsort`ing a `random` array tends to be a much faster way of generating the large arrays of random permutations we need. I was [advised against using this approach](https://github.com/scipy/scipy/pull/4824#discussion_r529271983). I agree with the sentiment to some extent, but all things considered, I have to disagree here because the "hackery" can be _so_ much better. If someone finds a case in which the hack is less performant in the future, they should just use `Generator`. I don't think we should need to Cythonize this.
- `Generator` has a `permuted` method that _does_ act on slices independently, so use that when it's available.

This PR uses these faster methods of generating arrays of permutations to improve performance substantially (25~50x).

```python3
import numpy as np
from scipy import stats

rng = np.random.default_rng()
rvs = rng.normal(size=1000)
def t_statistic(x, axis=-1):
    return stats.ttest_1samp(x, popmean=0, axis=axis).statistic

res = stats.permutation_test((rvs,), t_statistic, permutation_type='samples')
# PR: 671 ms ± 22.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# main: 17.9 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
res2 = stats.permutation_test((rvs,), t_statistic, permutation_type='samples',
                              random_state=rng)
# PR: 561 ms ± 11.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# main: 27.7 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

#### Additional information
- I might do a follow-up that implements this for the `independent` permutation type.
- I've written some (batched) Pythran exhaustive permutation generators for the exact tests. Eventually, we'll want to replace the pure-Python loops over `itertools.combinations`/`itertools.permutations` with those. Like `permuted`, these could permute the samples themselves rather than permuting indices and then doing fancy indexing. It would also avoid all the set difference stuff in `all_partitions`. It would save memory and improve speed if we took advantage of that.
- Eventually, we could improve  `_calculate_null_samples` by writing a standalone implementation (without having it call `_calculate_null_pairings`), but I found that didn't improve things very much unless `batch` is small and the number of observations per sample is quite large. Let's do this in chunks to keep reviews manageable.