@mdhaber 2022-11-06
> I could do more if you wish.

I'd appreciate it if you'd take a closer look. Maybe just confirming my timing comparison locally, reading through the changes enough to see how they work, and seeing whether you agree it is OK that this PR reverses my decision to take [this advice](https://github.com/scipy/scipy/pull/4824#discussion_r529271983). Thanks!